C#OOP/OOP-Exams/C# OOP Retake Exam - 19 December 2020/01. Structure_Skeleton/Entities/Characters/Warrior.cs
using System;
using System.Collections.Generic;
using System.Text;
using WarCroft.Constants;
using WarCroft.Entities.Characters.Contracts;
using WarCroft.Entities.Inventory;

namespace WarCroft.Entities.Characters
{
    public class Warrior : Character, IAttacker
    {
        public Warrior(string name) 
            : base(name, 100, 50, 40, new Satchel())
        {
        }

        public void Attack(Character character)
        {
            if (this.IsAlive && character.IsAlive)
            {
                if (this == character)
                {
                    throw new InvalidOperationException(ExceptionMessages.CharacterAttacksSelf);
                }
                character.TakeDamage(this.AbilityPoints);

                if (character.Health <= 0)
                {
                    character.IsAlive = false;
                }
            }
            else
            {
                throw new InvalidOperationException($"Must be alive to perform this action!");
            }
        }
    }
}
